+++ x/aswap/model.go
@@ -10,14 +10,6 @@ func init() {
 	migration.MustRegister(1, &Swap{}, migration.NoModification)
 }
 
-const (
-	// BucketName is where we store the swaps
-	BucketName = "swap"
-
-	// SequenceName is an auto-increment ID counter for swaps
-	SequenceName = "id"
-)
-
 var _ orm.CloneableData = (*Swap)(nil)
 
 // Validate ensures the Swap is valid
@@ -47,6 +39,9 @@ func (s *Swap) Validate() error {
 	if len(s.Memo) > maxMemoSize {
 		return errors.Wrapf(errors.ErrInput, "memo %s", s.Memo)
 	}
+	if err := s.Address.Validate(); err != nil {
+		return errors.Wrap(err, "address")
+	}
 	return nil
 }
 
@@ -59,6 +54,7 @@ func (s *Swap) Copy() orm.CloneableData {
 		Destination:  s.Destination,
 		Timeout:      s.Timeout,
 		Memo:         s.Memo,
+		Address:      s.Address.Clone(),
 	}
 }
 
@@ -72,28 +68,19 @@ func AsSwap(obj orm.Object) *Swap {
 	return obj.Value().(*Swap)
 }
 
-// Bucket is a type-safe wrapper around orm.Bucket
-type Bucket struct {
-	orm.IDGenBucket
+func NewBucket() orm.ModelBucket {
+	b := orm.NewModelBucket("swap", &Swap{},
+		orm.WithIDSequence(swapSeq),
+		orm.WithIndex("source", idxSource, false),
+		orm.WithIndex("destination", idxDestination, false),
+		orm.WithIndex("preimage_hash", idxPrehash, false),
+	)
+	return migration.NewModelBucket("aswap", b)
 }
 
-// NewBucket initializes a Bucket with default name
-//
-// inherit Get and Save from orm.Bucket
-// add Create
-func NewBucket() Bucket {
-	bucket := migration.NewBucket("aswap", BucketName,
-		orm.NewSimpleObj(nil, &Swap{})).
-		WithIndex("source", idxSource, false).
-		WithIndex("destination", idxDestination, false).
-		WithIndex("preimage_hash", idxPrehash, false)
+var swapSeq = orm.NewSequence("aswap", "id")
 
-	return Bucket{
-		IDGenBucket: orm.WithSeqIDGenerator(bucket, SequenceName),
-	}
-}
-
-func getSwap(obj orm.Object) (*Swap, error) {
+func toSwap(obj orm.Object) (*Swap, error) {
 	if obj == nil {
 		return nil, errors.Wrap(errors.ErrHuman, "Cannot take index of nil")
 	}
@@ -105,7 +92,7 @@ func getSwap(obj orm.Object) (*Swap, error) {
 }
 
 func idxSource(obj orm.Object) ([]byte, error) {
-	swp, err := getSwap(obj)
+	swp, err := toSwap(obj)
 	if err != nil {
 		return nil, err
 	}
@@ -113,7 +100,7 @@ func idxSource(obj orm.Object) ([]byte, error) {
 }
 
 func idxDestination(obj orm.Object) ([]byte, error) {
-	swp, err := getSwap(obj)
+	swp, err := toSwap(obj)
 	if err != nil {
 		return nil, err
 	}
@@ -121,7 +108,7 @@ func idxDestination(obj orm.Object) ([]byte, error) {
 }
 
 func idxPrehash(obj orm.Object) ([]byte, error) {
-	swp, err := getSwap(obj)
+	swp, err := toSwap(obj)
 	if err != nil {
 		return nil, err
 	}