+++ x/escrow/handler.go
@@ -38,7 +38,7 @@ func RegisterQuery(qr weave.QueryRouter) {
 // CreateEscrowHandler will set a name for objects in this bucket
 type CreateEscrowHandler struct {
 	auth   x.Authenticator
-	bucket Bucket
+	bucket orm.ModelBucket
 	bank   cash.CoinMover
 }
 
@@ -58,8 +58,8 @@ func (h CreateEscrowHandler) Check(ctx weave.Context, db weave.KVStore, tx weave
 	return res, nil
 }
 
-// Deliver moves the tokens from source to the escrow account if
-// all preconditions are met.
+// Deliver moves the tokens from source to the escrow account if all
+// preconditions are met.
 func (h CreateEscrowHandler) Deliver(ctx weave.Context, db weave.KVStore, tx weave.Tx) (*weave.DeliverResult, error) {
 	msg, err := h.validate(ctx, db, tx)
 	if err != nil {
@@ -72,6 +72,11 @@ func (h CreateEscrowHandler) Deliver(ctx weave.Context, db weave.KVStore, tx wea
 		source = x.MainSigner(ctx, h.auth).Address()
 	}
 
+	key, err := escrowSeq.NextVal(db)
+	if err != nil {
+		return nil, errors.Wrap(err, "cannot acquire key")
+	}
+
 	// create an escrow object
 	escrow := &Escrow{
 		Metadata:    &weave.Metadata{},
@@ -80,23 +85,17 @@ func (h CreateEscrowHandler) Deliver(ctx weave.Context, db weave.KVStore, tx wea
 		Destination: msg.Destination,
 		Timeout:     msg.Timeout,
 		Memo:        msg.Memo,
+		Address:     Condition(key).Address(),
 	}
-	obj, err := h.bucket.Build(db, escrow)
-	if err != nil {
-		return nil, err
+	if _, err := h.bucket.Put(db, key, escrow); err != nil {
+		return nil, errors.Wrap(err, "cannot store escrow")
 	}
 
-	// deposit amounts
-	escrowAddr := Condition(obj.Key()).Address()
-	sourceAddr := weave.Address(escrow.Source)
-	if err := cash.MoveCoins(db, h.bank, sourceAddr, escrowAddr, msg.Amount); err != nil {
+	// Deposit to the escrow account.
+	if err := cash.MoveCoins(db, h.bank, escrow.Source, escrow.Address, msg.Amount); err != nil {
 		return nil, err
 	}
-	// return id of escrow to use in future calls
-	res := &weave.DeliverResult{
-		Data: obj.Key(),
-	}
-	return res, h.bucket.Save(db, obj)
+	return &weave.DeliverResult{Data: key}, nil
 }
 
 // validate does all common pre-processing between Check and Deliver.
@@ -123,7 +122,7 @@ func (h CreateEscrowHandler) validate(ctx weave.Context, db weave.KVStore, tx we
 // ReleaseEscrowHandler will set a name for objects in this bucket.
 type ReleaseEscrowHandler struct {
 	auth   x.Authenticator
-	bucket Bucket
+	bucket orm.ModelBucket
 	bank   cash.Controller
 }
 
@@ -150,10 +149,8 @@ func (h ReleaseEscrowHandler) Deliver(ctx weave.Context, db weave.KVStore, tx we
 
 	// use amount in message, or
 	request := coin.Coins(msg.Amount)
-	key := msg.EscrowId
-	escrowAddr := Condition(key).Address()
 	if len(request) == 0 {
-		available, err := h.bank.Balance(db, escrowAddr)
+		available, err := h.bank.Balance(db, escrow.Address)
 		if err != nil {
 			return nil, err
 		}
@@ -161,19 +158,19 @@ func (h ReleaseEscrowHandler) Deliver(ctx weave.Context, db weave.KVStore, tx we
 	}
 
 	// withdraw the money from escrow to recipient
-	if err := cash.MoveCoins(db, h.bank, escrowAddr, escrow.Destination, request); err != nil {
+	if err := cash.MoveCoins(db, h.bank, escrow.Address, escrow.Destination, request); err != nil {
 		return nil, err
 	}
 
-	remainingCoins, err := h.bank.Balance(db, escrowAddr)
+	remainingCoins, err := h.bank.Balance(db, escrow.Address)
 	if err != nil {
 		return nil, err
 	}
 	if remainingCoins.IsPositive() {
-		return &weave.DeliverResult{Data: key}, nil
+		return &weave.DeliverResult{Data: msg.EscrowId}, nil
 	}
 	// Delete escrow when empty.
-	if err := h.bucket.Delete(db, key); err != nil {
+	if err := h.bucket.Delete(db, msg.EscrowId); err != nil {
 		return nil, err
 	}
 	return &weave.DeliverResult{}, nil
@@ -185,9 +182,9 @@ func (h ReleaseEscrowHandler) validate(ctx weave.Context, db weave.KVStore, tx w
 	if err := weave.LoadMsg(tx, &msg); err != nil {
 		return nil, nil, errors.Wrap(err, "load msg")
 	}
-	escrow, err := loadEscrow(h.bucket, db, msg.EscrowId)
-	if err != nil {
-		return nil, nil, err
+	var escrow Escrow
+	if err := h.bucket.One(db, msg.EscrowId, &escrow); err != nil {
+		return nil, nil, errors.Wrap(err, "cannot load escrow from the store")
 	}
 
 	// Arbiter or source must authorize this.
@@ -200,13 +197,13 @@ func (h ReleaseEscrowHandler) validate(ctx weave.Context, db weave.KVStore, tx w
 		return nil, nil, err
 	}
 
-	return &msg, escrow, nil
+	return &msg, &escrow, nil
 }
 
 // ReturnEscrowHandler will set a name for objects in this bucket
 type ReturnEscrowHandler struct {
 	auth   x.Authenticator
-	bucket Bucket
+	bucket orm.ModelBucket
 	bank   cash.Controller
 }
 
@@ -231,15 +228,14 @@ func (h ReturnEscrowHandler) Deliver(ctx weave.Context, db weave.KVStore, tx wea
 		return nil, err
 	}
 
-	escrowAddr := Condition(key).Address()
-	available, err := h.bank.Balance(db, escrowAddr)
+	available, err := h.bank.Balance(db, escrow.Address)
 	if err != nil {
 		return nil, err
 	}
 
 	// withdraw all coins from escrow to the defined "source"
 	dest := weave.Address(escrow.Source)
-	if err := cash.MoveCoins(db, h.bank, escrowAddr, dest, available); err != nil {
+	if err := cash.MoveCoins(db, h.bank, escrow.Address, dest, available); err != nil {
 		return nil, err
 	}
 	if err := h.bucket.Delete(db, key); err != nil {
@@ -255,22 +251,22 @@ func (h ReturnEscrowHandler) validate(ctx weave.Context, db weave.KVStore, tx we
 		return nil, nil, errors.Wrap(err, "load msg")
 	}
 
-	escrow, err := loadEscrow(h.bucket, db, msg.GetEscrowId())
-	if err != nil {
-		return nil, nil, err
+	var escrow Escrow
+	if err := h.bucket.One(db, msg.EscrowId, &escrow); err != nil {
+		return nil, nil, errors.Wrap(err, "cannot load escrow from the store")
 	}
 
 	if !weave.IsExpired(ctx, escrow.Timeout) {
 		return nil, nil, errors.Wrapf(errors.ErrState, "escrow not expired %v", escrow.Timeout)
 	}
 
-	return msg.EscrowId, escrow, nil
+	return msg.EscrowId, &escrow, nil
 }
 
 // UpdateEscrowHandler will set a name for objects in this bucket.
 type UpdateEscrowHandler struct {
 	auth   x.Authenticator
-	bucket Bucket
+	bucket orm.ModelBucket
 }
 
 var _ weave.Handler = UpdateEscrowHandler{}
@@ -294,7 +290,7 @@ func (h UpdateEscrowHandler) Deliver(ctx weave.Context, db weave.KVStore, tx wea
 		return nil, err
 	}
 
-	// update the escrow with message values
+	// Update the escrow with message values.
 	if msg.Source != nil {
 		escrow.Source = msg.Source
 	}
@@ -305,11 +301,9 @@ func (h UpdateEscrowHandler) Deliver(ctx weave.Context, db weave.KVStore, tx wea
 		escrow.Arbiter = msg.Arbiter
 	}
 
-	// save the updated escrow
-	key := msg.EscrowId
-	obj := orm.NewSimpleObj(key, escrow)
-	if err := h.bucket.Save(db, obj); err != nil {
-		return nil, err
+	// Save the updated escrow.
+	if _, err := h.bucket.Put(db, msg.EscrowId, escrow); err != nil {
+		return nil, errors.Wrap(err, "cannot save")
 	}
 	return &weave.DeliverResult{}, nil
 }
@@ -321,9 +315,9 @@ func (h UpdateEscrowHandler) validate(ctx weave.Context, db weave.KVStore, tx we
 		return nil, nil, errors.Wrap(err, "load msg")
 	}
 
-	escrow, err := loadEscrow(h.bucket, db, msg.GetEscrowId())
-	if err != nil {
-		return nil, nil, err
+	var escrow Escrow
+	if err := h.bucket.One(db, msg.EscrowId, &escrow); err != nil {
+		return nil, nil, errors.Wrap(err, "cannot load escrow from the store")
 	}
 
 	if weave.IsExpired(ctx, escrow.Timeout) {
@@ -348,18 +342,5 @@ func (h UpdateEscrowHandler) validate(ctx weave.Context, db weave.KVStore, tx we
 		}
 	}
 
-	return &msg, escrow, nil
-}
-
-// loadEscrow loads escrow and cast it, returns error if not present.
-func loadEscrow(bucket Bucket, db weave.KVStore, escrowID []byte) (*Escrow, error) {
-	obj, err := bucket.Get(db, escrowID)
-	if err != nil {
-		return nil, err
-	}
-	escrow := AsEscrow(obj)
-	if escrow == nil {
-		return nil, errors.Wrapf(errors.ErrEmpty, "escrow %d", escrowID)
-	}
-	return escrow, nil
+	return &msg, &escrow, nil
 }