+++ x/multisig/model.go
@@ -12,11 +12,6 @@ func init() {
 }
 
 const (
-	// BucketName is where we store the contracts
-	BucketName = "contracts"
-	// SequenceName is an auto-increment ID counter for contracts
-	SequenceName = "id"
-
 	// Maximum value a weight value can be set to. This is uint8 capacity
 	// but because we use protobuf for serialization, weight is represented
 	// by uint32 and we must manually force the limit.
@@ -50,6 +45,9 @@ func (c *Contract) Validate() error {
 	case n > maxParticipantsAllowed:
 		return errors.Wrap(errors.ErrModel, "too many participants")
 	}
+	if err := c.Address.Validate(); err != nil {
+		return errors.Wrap(err, "address")
+	}
 	return validateWeights(errors.ErrModel,
 		c.Participants, c.ActivationThreshold, c.AdminThreshold)
 }
@@ -69,60 +67,15 @@ func (c *Contract) Copy() orm.CloneableData {
 		Participants:        ps,
 		ActivationThreshold: c.ActivationThreshold,
 		AdminThreshold:      c.AdminThreshold,
+		Address:             c.Address.Clone(),
 	}
 }
 
-// ContractBucket is a type-safe wrapper around orm.Bucket
-type ContractBucket struct {
-	orm.Bucket
-	idSeq orm.Sequence
+func NewContractBucket() orm.ModelBucket {
+	b := orm.NewModelBucket("contracts", &Contract{},
+		orm.WithIDSequence(contractSeq),
+	)
+	return migration.NewModelBucket("multisig", b)
 }
 
-// NewContractBucket initializes a ContractBucket with default name
-//
-// inherit Get and Save from orm.Bucket
-// add run-time check on Save
-func NewContractBucket() ContractBucket {
-	bucket := migration.NewBucket("multisig", BucketName,
-		orm.NewSimpleObj(nil, new(Contract)))
-	return ContractBucket{
-		Bucket: bucket,
-		idSeq:  bucket.Sequence(SequenceName),
-	}
-}
-
-// Save enforces the proper type
-func (b ContractBucket) Save(db weave.KVStore, obj orm.Object) error {
-	if _, ok := obj.Value().(*Contract); !ok {
-		return errors.WithType(errors.ErrModel, obj.Value())
-	}
-	return b.Bucket.Save(db, obj)
-}
-
-// Build assigns an ID to given contract instance and returns it as an orm
-// Object. It does not persist the escrow in the store.
-func (b ContractBucket) Build(db weave.KVStore, c *Contract) (orm.Object, error) {
-	key, err := b.idSeq.NextVal(db)
-	if err != nil {
-		return nil, err
-	}
-	return orm.NewSimpleObj(key, c), nil
-}
-
-// GetContract returns a contract with given ID.
-func (b ContractBucket) GetContract(store weave.KVStore, contractID []byte) (*Contract, error) {
-	obj, err := b.Get(store, contractID)
-	if err != nil {
-		return nil, errors.Wrap(err, "bucket lookup")
-	}
-
-	if obj == nil || obj.Value() == nil {
-		return nil, errors.Wrapf(errors.ErrNotFound, "contract id %q", contractID)
-	}
-
-	c, ok := obj.Value().(*Contract)
-	if !ok {
-		return nil, errors.Wrapf(errors.ErrModel, "invalid type: %T", obj.Value())
-	}
-	return c, nil
-}
+var contractSeq = orm.NewSequence("contracts", "id")